alpha 0.0.1
After a LONG time of boning up on Java and Processing syntax (I swear it wasn't this complicated before, GP Blocks must have actively asphyxiated my brain somehow), I finally have a "working" GUI ready to show. That was uncomfortably long, but it's a massive jump in syntax. After this, we will have more concrete progress. Current features include a rotating cube, that you can click and rotate by clicking and dragging. Double-click to reset the viewing angle (honestly useless at this point, but eventually it will have a good use). It is a perspective camera view because LIBRARIES EXIST. I know that I swore I wouldn't use anything other than an isomorphic camera viewing angle, but it comes with the GUI creator tool.

GUI Elements:

The top left is the ViewFinder. The name is explanatory, it shows you what you are working on. The top right is the MessageBoard. It says W.I.P. because it obviously is not functional. It will have the current conversation displayed as a log of messages. I am trying to find a good color scheme for that one. Below the ViewFinder is a small strip, which will show the current filename. Note that you can select it to write text. That's intentional. Bottom Left is our familiar TextTron, and it actually has the >: prompt I couldn't figure out how to implement in GP Blocks. No, It isn't part of TextTron, that is just a label. Bottom right is a NEW idea I had off the top of my head, say hello to leReporter (sic)! It acts as a standard output feed for the user to understand why stuff didn't work or to answer queries that might be implemented. I will probably be adding TelegramBot to the MessageBoard panel, because having responsive output is essential, and there is nowhere else to put it currently. Finally, SEVEN NEW BUTTONS (in not-so-subtle rainbow colors) are visible in the center of the screen! These will do what they say, making a new file, saving the current file (I intend to use the FileNameShower to allow you to save it to a new filename), loading another filename (this I think will require a pop-up to select the appropriate file), undoing your last transformative action, exporting the file to a form that can be used in Slic3r, logging out, and logging in. They just... don't do it now. Sorry, trying to hunt down the appropriate libraries and learning the syntax of their contents took way longer than I anticipated.

YOU WILL NEED PROCESSING TO RUN THIS PROGRAM. Probably processing 4. You will also need the G4P_GUI package and the PeasyCam library that it comes with.