New update: Not much visually
I have added a couple of files, the main backbone of the physics system we will be using.

P3d(insert subclass here)!

The vanilla PShape is horribly clunky and tedious to use for various 3d shapes. Every Vertex has to be listed in one contiguous stream, which means you need to list 16 links to get all 12 edges of a cube. UNACCEPTABLE!!! There's got to be a quicker way...

So I am introducing the P3dShape Class! It is made of the P3dFace, P3dLine, and P3dPoint classes, which are able to coexist with NURBS functionality!

That's right, WE ARE GETTING TUBULAR! Curved surfaces are a really important thing in manufacturing, for many applications. Car hoods, furniture design, and even coffee mugs (the ergonomic kind, that is) utilize computationally generated curves. Yes, we might be making Java mugs, with Java Language. The jokes write themselves it seems.

P3dShape is not implemented yet. That is the next big hurdle, and I will need to refine the other new classes to make sure they function appropriately. Can't have a face not meeting the others on a solid object, that does not make any physical or geometrical sense.

Uses NURBS library, so if that goes down, we might have a problem unless we can get a local copy... I don't know if that is allowed or not.